pyproject.toml
[project]
name = "strategy_platform_service"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic-settings>=2.0",
    "structlog>=24.0",
    "fastapi>=0.115.0",
    "sqlalchemy>=2.0",
    "asyncpg>=0.31.0",
    "psycopg2-binary>=2.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "alembic>=1.13.0",
    "redis[hiredis]>=5.0.0",
    "celery[redis]>=5.3.0",
    "sqladmin>=0.18.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "itsdangerous>=2.2.0",
    "freqtrade>=2026.2",
    "greenlet>=3.3.2",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "mypy>=1.0",
    "ruff>=0.8",
    "pre-commit>=3.0",
    "pytest-asyncio>=1.3.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pyright]
pythonVersion = "3.11"
extraPaths = ["."]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=0"
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"
